Reorganized the ACTIONX implementation - made large parts private
