+++ opm/parser/eclipse/EclipseState/Schedule/Action/Actions.hpp
@@ -25,7 +25,7 @@
 #include <ctime>
 #include <map>
 
-#include <opm/parser/eclipse/EclipseState/Schedule/ActionX.hpp>
+#include <opm/parser/eclipse/EclipseState/Schedule/Action/ActionX.hpp>
 
 namespace Opm {
