+++ src/opm/parser/eclipse/EclipseState/Schedule/Action/ActionValue.hpp
@@ -1,0 +1,45 @@
+#ifndef ACTION_VALUE_HPP
+#define ACTION_VALUE_HPP
+
+#include "WellSet.hpp"
+
+enum TokenType {
+  number,        //  0
+  ecl_expr,      //  1
+  open_paren,    //  2
+  close_paren,   //  3
+  op_gt,         //  4
+  op_ge,         //  5
+  op_lt,         //  6
+  op_le,         //  7
+  op_eq,         //  8
+  op_ne,         //  9
+  op_and,        // 10
+  op_or,         // 11
+  end,           // 12
+  error          // 13
+};
+
+
+
+namespace Opm {
+
+class ActionValue {
+public:
+    explicit ActionValue(double value);
+    ActionValue() = default;
+
+    bool eval_cmp(TokenType op, const ActionValue& rhs, WellSet& matching_wells) const;
+    void add_well(const std::string& well, double value);
+    double scalar() const;
+
+private:
+    bool eval_cmp_wells(TokenType op, double rhs, WellSet& matching_wells) const;
+
+    double scalar_value;
+    double is_scalar = false;
+    std::vector<std::pair<std::string, double>> well_values;
+};
+
+}
+#endif
