+++ src/opm/parser/eclipse/EclipseState/Schedule/Action/ActionX.cpp
@@ -18,7 +18,7 @@
  */
 
 #include <opm/parser/eclipse/Deck/DeckKeyword.hpp>
-#include <opm/parser/eclipse/EclipseState/Schedule/ActionX.hpp>
+#include <opm/parser/eclipse/EclipseState/Schedule/Action/ActionX.hpp>
 
 
 namespace Opm {
